requirements.txt
imap-tools
loguru
pydantic
scikit-learn
pandas
python-dotenv
requests
unidecode
